Fix #7742, 66dd7c3: widget position can be negative

+++ src/widget.cpp
@@ -2494,7 +2494,7 @@ void NWidgetLeaf::Draw(const Window *w)
 
 		case WWT_RESIZEBOX:
 			assert(this->widget_data == 0);
-			DrawResizeBox(r, this->colour, this->pos_x < (uint)(w->width / 2), !!(w->flags & WF_SIZING));
+			DrawResizeBox(r, this->colour, this->pos_x < (w->width / 2), !!(w->flags & WF_SIZING));
 			break;
 
 		case WWT_CLOSEBOX:
